[FIX] 위,경도 타입을 BigDecimal로 변경해 DECIMAL(10,7) 스키마와 일치

+++ firm-server/src/main/java/chill_logistics/firm_server/domain/entity/Firm.java
@@ -5,6 +5,7 @@ import lib.entity.BaseEntity;
 import lombok.Getter;
 import org.hibernate.annotations.GenericGenerator;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 @Getter
@@ -65,8 +66,8 @@ public class Firm extends BaseEntity {
     private String fullAddress;
 
     @Column(name = "latitude", nullable = false, precision = 10, scale = 7)
-    private double latitude;
+    private BigDecimal latitude;
 
     @Column(name = "longitude", nullable = false, precision = 10, scale = 7)
-    private double longitude;
+    private BigDecimal longitude;
 }
